+++ src/components/slide/content/feed-selector.jsx
@@ -12,7 +12,8 @@ import MultiSelectComponent from "../../util/forms/multiselect-dropdown/multi-dr
 import idFromUrl from "../../util/helpers/id-from-url";
 import ContentForm from "./content-form";
 import MultiselectFromEndpoint from "./multiselect-from-endpoint";
-import PosterSelector from "./poster-selector";
+import PosterSelectorV1 from "./poster/poster-selector-v1";
+import PosterSelectorV2 from "./poster/poster-selector-v2";
 
 /**
  * Feed selector.
@@ -96,9 +97,19 @@ function FeedSelector({
     onChange(newValue);
   };
 
-  const configurationChange = ({ target }) => {
+  const configurationChange = ({ target = null, targets = null }) => {
     const configuration = { ...value.configuration };
-    set(configuration, target.id, target.value);
+
+    if (target !== null) {
+      set(configuration, target.id, target.value);
+    }
+
+    if (targets !== null) {
+      targets.forEach(({ id, value: targetValue }) => {
+        set(configuration, id, targetValue);
+      });
+    }
+
     const newValue = { ...value, configuration };
     onChange(newValue);
   };
@@ -129,10 +140,21 @@ function FeedSelector({
     }
     if (element?.input === "poster-selector") {
       return (
-        <PosterSelector
+        <PosterSelectorV1
           key={element.key}
           feedSource={feedSourceData}
           configurationChange={configurationChange}
+          getValueFromConfiguration={getValueFromConfiguration}
+        />
+      );
+    }
+    if (element?.input === "poster-selector-v2") {
+      return (
+        <PosterSelectorV2
+          key={element.key}
+          feedSource={feedSourceData}
+          configurationChange={configurationChange}
+          configuration={value.configuration}
           getValueFromConfiguration={getValueFromConfiguration}
         />
       );